+++ src/commands/json/cmd-json.mts
@@ -1,12 +1,8 @@
-import fs from 'node:fs'
 import path from 'node:path'
 
-import { logger } from '@socketsecurity/registry/lib/logger'
-
-import constants from '../../constants.mts'
+import { handleCmdJson } from './handle-cmd-json.mts'
 import { commonFlags } from '../../flags.mts'
 import { meowOrExit } from '../../utils/meow-with-subcommands.mts'
-import { tildify } from '../../utils/tildify.mts'
 
 import type { CliCommandConfig } from '../../utils/meow-with-subcommands.mts'
 
@@ -50,28 +46,5 @@ async function run(
   // If given path is absolute then cwd should not affect it.
   cwd = path.resolve(process.cwd(), cwd)
 
-  logger.info('Target cwd:', constants.ENV.VITEST ? '<redacted>' : tildify(cwd))
-
-  const sjpath = path.join(cwd, 'socket.json')
-  const tildeSjpath = constants.ENV.VITEST ? '<redacted>' : tildify(sjpath)
-
-  if (!fs.existsSync(sjpath)) {
-    logger.fail(`Not found: ${tildeSjpath}`)
-    process.exitCode = 1
-    return
-  }
-
-  if (!fs.lstatSync(sjpath).isFile()) {
-    logger.fail(
-      `This is not a regular file (maybe a directory?): ${tildeSjpath}`,
-    )
-    process.exitCode = 1
-    return
-  }
-
-  const data = fs.readFileSync(sjpath, 'utf8')
-
-  logger.success(`This is the contents of ${tildeSjpath}:`)
-  logger.error('')
-  logger.log(data)
+  await handleCmdJson(cwd)
 }
