+++ src/commands/json/handle-cmd-json.mts
@@ -1,0 +1,5 @@
+import { outputCmdJson } from './output-cmd-json.mts'
+
+export async function handleCmdJson(cwd: string) {
+  await outputCmdJson(cwd)
+}
